updated error handling in notiications and upgraded notification funtionality to seperate app
